fix: apply pr fixes
