+++ docs/api/openAPI.yaml
@@ -61,7 +61,7 @@ components:
       properties:
         deliveries:
           items:
-            $ref: '#/components/schemas/DeliveryAnonymized'
+            $ref: '#/components/schemas/ReportedAnonymizedDelivery'
           maxItems: 50
           type: array
           uniqueItems: true
@@ -112,7 +112,7 @@ components:
           type: string
         productions:
           items:
-            $ref: '#/components/schemas/AllocatedPlannedProductionOutputAnonymized'
+            $ref: '#/components/schemas/ReportedAnonymizedProduction'
           maxItems: 50
           type: array
           uniqueItems: true
@@ -121,7 +121,7 @@ components:
           type: number
         stocks:
           items:
-            $ref: '#/components/schemas/AllocatedStockAnonymized'
+            $ref: '#/components/schemas/ReportedAnonymizedStock'
           maxItems: 50
           type: array
           uniqueItems: true
@@ -1655,6 +1655,240 @@ components:
       - date
       - daysOfSupply
       type: object
+    ReportedAnonymizedDelivery:
+      additionalProperties: false
+      properties:
+        aggregatedMaterialDataNodeId:
+          format: uuid
+          maxItems: 50
+          type: string
+        arrivalType:
+          enum:
+          - estimated-departure
+          - actual-departure
+          - estimated-arrival
+          - actual-arrival
+          maxItems: 50
+          type: string
+        dateOfArrival:
+          format: date-time
+          maxItems: 50
+          type: string
+        dateOfDeparture:
+          format: date-time
+          maxItems: 50
+          type: string
+        departureType:
+          enum:
+          - estimated-departure
+          - actual-departure
+          - estimated-arrival
+          - actual-arrival
+          maxItems: 50
+          type: string
+        destinationBpnsAnonymized:
+          maxItems: 50
+          type: string
+        lastUpdatedOnDateTime:
+          format: date-time
+          maxItems: 50
+          type: string
+        measurementUnit:
+          enum:
+          - unit:piece
+          - unit:set
+          - unit:pair
+          - unit:page
+          - unit:cycle
+          - unit:kilowattHour
+          - unit:gram
+          - unit:kilogram
+          - unit:tonneMetricTon
+          - unit:tonUsOrShortTonUkorus
+          - unit:ounceAvoirdupois
+          - unit:pound
+          - unit:metre
+          - unit:centimetre
+          - unit:kilometre
+          - unit:inch
+          - unit:foot
+          - unit:yard
+          - unit:squareCentimetre
+          - unit:squareMetre
+          - unit:squareInch
+          - unit:squareFoot
+          - unit:squareYard
+          - unit:cubicCentimetre
+          - unit:cubicMetre
+          - unit:cubicInch
+          - unit:cubicFoot
+          - unit:cubicYard
+          - unit:litre
+          - unit:millilitre
+          - unit:hectolitre
+          - unit:secondUnitOfTime
+          - unit:minuteUnitOfTime
+          - unit:hourUnitOfTime
+          - unit:day
+          maxItems: 50
+          type: string
+        originBpnsAnonymized:
+          maxItems: 50
+          type: string
+        quantity:
+          format: double
+          type: number
+        uuid:
+          format: uuid
+          maxItems: 50
+          type: string
+      required:
+      - dateOfDeparture
+      - departureType
+      - destinationBpnsAnonymized
+      - lastUpdatedOnDateTime
+      - originBpnsAnonymized
+      type: object
+    ReportedAnonymizedProduction:
+      additionalProperties: false
+      properties:
+        aggregatedMaterialDataNodeId:
+          format: uuid
+          maxItems: 50
+          type: string
+        estimatedTimeOfCompletion:
+          format: date-time
+          maxItems: 50
+          type: string
+        lastUpdatedOnDateTime:
+          format: date-time
+          maxItems: 50
+          type: string
+        materialGlobalAssetIdAnonymized:
+          maxItems: 50
+          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
+          type: string
+        measurementUnit:
+          enum:
+          - unit:piece
+          - unit:set
+          - unit:pair
+          - unit:page
+          - unit:cycle
+          - unit:kilowattHour
+          - unit:gram
+          - unit:kilogram
+          - unit:tonneMetricTon
+          - unit:tonUsOrShortTonUkorus
+          - unit:ounceAvoirdupois
+          - unit:pound
+          - unit:metre
+          - unit:centimetre
+          - unit:kilometre
+          - unit:inch
+          - unit:foot
+          - unit:yard
+          - unit:squareCentimetre
+          - unit:squareMetre
+          - unit:squareInch
+          - unit:squareFoot
+          - unit:squareYard
+          - unit:cubicCentimetre
+          - unit:cubicMetre
+          - unit:cubicInch
+          - unit:cubicFoot
+          - unit:cubicYard
+          - unit:litre
+          - unit:millilitre
+          - unit:hectolitre
+          - unit:secondUnitOfTime
+          - unit:minuteUnitOfTime
+          - unit:hourUnitOfTime
+          - unit:day
+          maxItems: 50
+          type: string
+        productionSiteBpnsAnonymized:
+          maxItems: 50
+          type: string
+        quantity:
+          format: double
+          type: number
+        uuid:
+          format: uuid
+          maxItems: 50
+          type: string
+      required:
+      - estimatedTimeOfCompletion
+      - lastUpdatedOnDateTime
+      - productionSiteBpnsAnonymized
+      type: object
+    ReportedAnonymizedStock:
+      additionalProperties: false
+      properties:
+        aggregatedMaterialDataNodeId:
+          format: uuid
+          maxItems: 50
+          type: string
+        blocked:
+          type: boolean
+        lastUpdatedOnDateTime:
+          format: date-time
+          maxItems: 50
+          type: string
+        measurementUnit:
+          enum:
+          - unit:piece
+          - unit:set
+          - unit:pair
+          - unit:page
+          - unit:cycle
+          - unit:kilowattHour
+          - unit:gram
+          - unit:kilogram
+          - unit:tonneMetricTon
+          - unit:tonUsOrShortTonUkorus
+          - unit:ounceAvoirdupois
+          - unit:pound
+          - unit:metre
+          - unit:centimetre
+          - unit:kilometre
+          - unit:inch
+          - unit:foot
+          - unit:yard
+          - unit:squareCentimetre
+          - unit:squareMetre
+          - unit:squareInch
+          - unit:squareFoot
+          - unit:squareYard
+          - unit:cubicCentimetre
+          - unit:cubicMetre
+          - unit:cubicInch
+          - unit:cubicFoot
+          - unit:cubicYard
+          - unit:litre
+          - unit:millilitre
+          - unit:hectolitre
+          - unit:secondUnitOfTime
+          - unit:minuteUnitOfTime
+          - unit:hourUnitOfTime
+          - unit:day
+          maxItems: 50
+          type: string
+        quantity:
+          format: double
+          type: number
+        stockLocationBpnsAnonymized:
+          maxItems: 50
+          type: string
+        uuid:
+          format: uuid
+          maxItems: 50
+          type: string
+      required:
+      - lastUpdatedOnDateTime
+      - measurementUnit
+      - stockLocationBpnsAnonymized
+      type: object
     ReportedMaterialStockDto:
       additionalProperties: false
       properties:
@@ -1971,7 +2205,7 @@ paths:
           description: Material does not exist.
       summary: Get all aggregated data for the given material
       tags:
-      - aggregated-material-data-node-controller
+      - aggregated-material-data-controller
   /batch/partner-data-update:
     get:
       operationId: history
